# src/main/resources/static/servicio-api-docs.yaml
openapi: 3.0.0
info:
  title: API Colaboradores recomendados
  description: |
    Esta API permite a las organizaciones interesadas obtener colaboradores recomendados para reconocer. Tomando en cuenta los puntos acumulados y el número de donaciones realizadas, se pueden obtener colaboradores que cumplan con ciertos criterios. A su vez, ofrece un endpoint para hacer la carga de colaboradores en la base de datos.
  version: 1.0.0
paths:
  /colaboradores:
    get:
      summary: Obtener colaboradores
      description: Recuperar una lista de colaboradores basada en criterios especificados
      parameters:
        - name: minPuntos
          in: query
          required: true
          schema:
            type: number
            format: double
          description: Puntos mínimos requeridos
        - name: minDonaciones
          in: query
          required: true
          schema:
            type: integer
          description: Número mínimo de donaciones requeridas
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Número de página para la paginación
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
          description: Número máximo de colaboradores a devolver
        - name: sort
          in: query
          required: false
          schema:
            type: string
            default: "puntos,desc;donaciones,desc"
          description: Criterios de ordenamiento
      responses:
        '200':
          description: Operación exitosa
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ColaboradorDTO'
    post:
      summary: Actualizar la base de datos de colaboradores
      description: Guardar o actualizar en bloque colaboradores en la base de datos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ColaboradorDTO'
      responses:
        '200':
          description: Operación exitosa
components:
  schemas:
    ColaboradorDTO:
      type: object
      required:
        - nro_documento
        - tipo_documento
        - nombre
        - apellido
        - medios_de_contacto
        - puntos
        - donaciones
      properties:
        nro_documento:
          type: string
        tipo_documento:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        medios_de_contacto:
          type: array
          items:
            $ref: '#/components/schemas/MedioDeContactoDTO'
        puntos:
          type: number
          format: double
        donaciones:
          type: integer
    MedioDeContactoDTO:
      type: object
      properties:
        tipo_medio_contacto:
          type: string
        detalle:
          type: string